--- app/templates/test.html
@@ -1,8 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Test Page{% endblock %}
-
-{% block content %}
-    <h1>Test Page</h1>
-    <p>{{ context.message }}</p>
-{% endblock %}
